Collapse a vector into a natural language string with a maximum number of elements

Collapse a vector into a natural language string with a maximum number of elements

plu_more(
  x,
  max = 5,
  type = TRUE,
  fn = NULL,
  ...,
  det = "more",
  sep = ", ",
  conj = " and ",
  oxford = getOption("plu.oxford_comma", FALSE)
)

more(
  x,
  max = 5,
  type = TRUE,
  fn = NULL,
  ...,
  det = "more",
  sep = ", ",
  conj = " and ",
  oxford = getOption("plu.oxford_comma", FALSE)
)

Arguments

x
A character vector (or a vector coercible to character).

max
The maximum number of items to list. Additional arguments are replaced with "n more". Defaults to 5.

type
A logical or character.

If a character, type is pasted after the number of elements.

If TRUE, the default, the first class of x is pasted after the number of elements.

If x is not atomic (e.g. a list), "element" is used in place of a class name.

If FALSE or NA, nothing is pasted after the number of elements.

fn
A function to apply to the number of additional elements. Default to NULL, which applies no function.

...
Additional arguments to fn.

det
A determiner to place before the number of additional elements. Defaults to "more".

sep
A character to place between list items. Defaults to ", "

conj
A character to place between the penultimate and last list items. Defaults to " and ". If NULL, sep is used.

oxford
A logical indicating whether to place sep before conj (x, y, and z) or not (x, y and z) in lists of length three or more. Defaults to FALSE. The default can be changed by setting options(plu.oxford_comma).

Value

A character vector of length 1.

Examples

plu::more(letters)

#> [1] "a, b, c, d, e and 21 more characters"

# Setting `max`
plu::more(letters, max = 10)

#> [1] "a, b, c, d, e, f, g, h, i, j and 16 more characters"

plu::more(letters, max = 27)

#> [1] "a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y and z"

# If `max` is Inf or NULL, all elements will be preserved
plu::more(letters, max = Inf)

#> [1] "a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y and z"

# If `max` is less than one, no elements will be preserved
plu::more(letters, max = 0)

#> [1] "26 characters"

# Setting `type`
plu::more(letters, type = "letter")

#> [1] "a, b, c, d, e and 21 more letters"

# If `type` is FALSE or NULL, no type will be included
plu::more(letters, type = FALSE)

#> [1] "a, b, c, d, e and 21 more"

# Automatically generating type
plu::more(1:100)

#> [1] "1, 2, 3, 4, 5 and 95 more integers"

plu::more(as.list(1:100))

#> [1] "1, 2, 3, 4, 5 and 95 more elements"

plu::more(fracture::fracture((1:9) / (9:1)))

#> [1] "1/9, 1/4, 3/7, 2/3, 1/1 and 4 more fractures"

# Applying a function to the number
plu::more(letters, fn = nombre::cardinal)

#> [1] "a, b, c, d, e and twenty-one more characters"

message(plu::more(sapply(letters, crayon::blue), fn = crayon::blue))

#> a, b, c, d, e and 21 more characters

# Automatic pluralization of type
fish <- c("sea bass", "pale chub", "crucian carp", "dace", "coelecanth")
plu::more(fish, max = 3, type = "fish")

#> [1] "sea bass, pale chub, crucian carp and 2 more fish"

plu::more(fish, max = 4, type = "fish")

#> [1] "sea bass, pale chub, crucian carp, dace and 1 more fish"

teeth <- c("incisor", "canine", "premolar", "molar", "wisdom tooth")
more(teeth, max = 3, type = "tooth")

#> [1] "incisor, canine, premolar and 2 more teeth"

more(teeth, max = 4, type = "tooth")

#> [1] "incisor, canine, premolar, molar and 1 more tooth"

cacti <- c("saguaro", "prickly pear", "barrel", "star", "chin")
more(cacti, max = 3, type = "cactus")

#> [1] "saguaro, prickly pear, barrel and 2 more cacti"

more(cacti, max = 4, type = "cactus")

#> [1] "saguaro, prickly pear, barrel, star and 1 more cactus"

# Formatting with `sep`, `conj`, and `oxford`
more(letters, conj = " or ", det = "other")

#> [1] "a, b, c, d, e or 21 other characters"

more(letters, oxford = FALSE)

#> [1] "a, b, c, d, e and 21 more characters"

more(letters, oxford = TRUE)

#> [1] "a, b, c, d, e, and 21 more characters"

more(letters, sep = "; ", oxford = TRUE)

#> [1] "a; b; c; d; e; and 21 more characters"

more(letters, sep = " + ", conj = " + ... + ")

#> [1] "a + b + c + d + e + ... + 21 more characters"

# Using plu_more() within a function
verbose_sqrt <- function(x) {
  if (any(x < 0)) {
    problems <- crayon::silver(encodeString(x[x < 0], quote = "`"))

    warning(
      "Square root is undefined for ",
      plu_more(problems, fn = crayon::silver, type = "input."),
      call. = FALSE
    )
  }

  suppressWarnings(sqrt(x))
}

verbose_sqrt(-10:10)

#> Warning: Square root is undefined for `-10`, `-9`, `-8`, `-7`, `-6` and 5 more inputs.

#>  [1]      NaN      NaN      NaN      NaN      NaN      NaN      NaN      NaN
#>  [9]      NaN      NaN 0.000000 1.000000 1.414214 1.732051 2.000000 2.236068
#> [17] 2.449490 2.645751 2.828427 3.000000 3.162278