Pluralize a word

Pluralize a word

plu_ralize(
  x,
  irregulars = getOption("plu.irregulars", c("moderate", "conservative", "liberal",
    "none"))
)

ralize(
  x,
  irregulars = getOption("plu.irregulars", c("moderate", "conservative", "liberal",
    "none"))
)

Arguments

x
A character vector of English words to be pluralized

irregulars
What level of irregularity to use in pluralization. "moderate" uses the most common pluralization. "conservative" uses the most common irregular plural if one exists, even if a regular plural is more common. "liberal" uses a regular plural if it exists, even if an irregular plural is more common. "none" attempts to apply regular noun pluralization rules to all words. Defaults to "moderate". The default can be changed by setting options(plu.irregulars). See examples.

Source

Irregular plurals list adapted from the Automatically Generated Inflection Database (AGID)

See plu-package for more details.

Value

The character vector x pluralized

See also

Examples

plu::ralize("word")

#> [1] "words"

plu::ralize(c("group", "word"))

#> [1] "groups" "words" 

plu::ralize(c("formula", "person", "child"), irregulars = "conservative")

#> [1] "formulae" "people"   "children"

plu::ralize(c("formula", "person", "child"), irregulars = "moderate")

#> [1] "formulas" "people"   "children"

plu::ralize(c("formula", "person", "child"), irregulars = "liberal")

#> [1] "formulas" "persons"  "children"

plu::ralize(c("formula", "person", "child"), irregulars = "none")

#> [1] "formulas" "persons"  "childs"